Refactor: Move BackInTimeEventData to model module
