+++ tooling/flipper-lib/src/commonMain/kotlin/com/kitakkun/backintime/tooling/flipper/FlipperAppStateOwner.kt
@@ -4,6 +4,7 @@ package com.kitakkun.backintime.tooling.flipper
 
 import com.kitakkun.backintime.core.websocket.event.BackInTimeDebugServiceEvent
 import com.kitakkun.backintime.core.websocket.event.BackInTimeDebuggerEvent
+import com.kitakkun.backintime.tooling.model.BackInTimeEventData
 import com.kitakkun.backintime.tooling.model.ClassInfo
 import com.kitakkun.backintime.tooling.model.DependencyInfo
 import com.kitakkun.backintime.tooling.model.InstanceInfo
